+++ tests/unit/session/test_clients.py
@@ -26,9 +26,7 @@ import google.cloud.bigquery_connection_v1
 import google.cloud.bigquery_storage_v1
 import google.cloud.functions_v2
 import google.cloud.resourcemanager_v3
-import pytest
 
-import bigframes.constants as constants
 import bigframes.session.clients as clients
 import bigframes.version
 
@@ -183,21 +181,3 @@ def test_user_agent_in_jupyter(monkeypatch):
 
     # We still need to include attribution to bigframes
     assert_clients_w_user_agent(provider, f"bigframes/{bigframes.version.__version__}")
-
-
-def test_clients_provider_no_location():
-    with pytest.raises(ValueError, match="Must set location to use regional endpoints"):
-        clients.ClientsProvider(use_regional_endpoints=True)
-
-
-@pytest.mark.parametrize(
-    "bigquery_location",
-    # Sort the set to avoid nondeterminism.
-    sorted(constants.REP_NOT_ENABLED_BIGQUERY_LOCATIONS),
-)
-def test_clients_provider_use_regional_endpoints_non_rep_locations(bigquery_location):
-    with pytest.raises(
-        ValueError,
-        match="Support for regional endpoints may not be available in the location",
-    ):
-        clients.ClientsProvider(location=bigquery_location, use_regional_endpoints=True)
